update to be more like creat-react-app@3.3

+++ bin/monkey-react-scripts.js
@@ -17,37 +17,31 @@ const scriptIndex = args.findIndex(
 const script = scriptIndex === -1 ? args[0] : args[scriptIndex];
 const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
 
-switch (script) {
-    case 'build':
-    case 'start':
-    case 'test': {
-        const result = spawn.sync(
-          'node',
-          nodeArgs
-            .concat(require.resolve('../scripts/' + script))
-            .concat(args.slice(scriptIndex + 1)),
-          { stdio: 'inherit' }
-        );
-        if (result.signal) {
-            if (result.signal === 'SIGKILL') {
-                console.log(
-                  'The build failed because the process exited too early. ' +
-                  'This probably means the system ran out of memory or someone called ' +
-                  '`kill -9` on the process.'
-                );
-            } else if (result.signal === 'SIGTERM') {
-                console.log(
-                  'The build failed because the process exited too early. ' +
-                  'Someone might have called `kill` or `killall`, or the system could ' +
-                  'be shutting down.'
-                );
-            }
-            process.exit(1);
+if (['build', 'start', 'test'].includes(script)) {
+    const result = spawn.sync(
+      'node',
+      nodeArgs
+        .concat(require.resolve('../scripts/' + script))
+        .concat(args.slice(scriptIndex + 1)),
+      { stdio: 'inherit' }
+    );
+    if (result.signal) {
+        if (result.signal === 'SIGKILL') {
+            console.log(
+              'The build failed because the process exited too early. ' +
+              'This probably means the system ran out of memory or someone called ' +
+              '`kill -9` on the process.'
+            );
+        } else if (result.signal === 'SIGTERM') {
+            console.log(
+              'The build failed because the process exited too early. ' +
+              'Someone might have called `kill` or `killall`, or the system could ' +
+              'be shutting down.'
+            );
         }
-        process.exit(result.status);
-        break;
+        process.exit(1);
     }
-    default:
-        console.log('Unknown script "' + script + '".');
-        break;
+    process.exit(result.status);
+} else {
+    console.log('Unknown script "' + script + '".');
 }
